sapphire/services/auth/requirements.txt
fastapi==0.118.3
uvicorn[standard]==0.37.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pydantic==2.12.0
pydantic-settings==2.11.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.21
httpx==0.28.0
email-validator==2.3.0